Reducing the usage of some global object variables
ParserOutput class nowadays (as of MW 1.34, but it was added way before that IIRC)
has an addModuleStyles() method so just use that.

One usage of $wgOut remains in Link#parseDescription and I don't really see a way around that
(other than changing it to RequestContext::getMain()->getOutput() but isn't that really just $wgOut in disguise?).

Change-Id: I6e27eea4148d3c1cd306066770dd9951741fdeb3