Fix bug with the Special:LinkEdit page displaying raw HTML instead of wikitext in the "description" box
It was because the special page uses Link#getLinkByPageID, which used to return parsed description (HTML)
instead of raw data stored in the link table in the DB (wikitext).
Changed the parsing to happen closer to the output instead and made getLinkByPageID() return the raw data
from the DB without any transformations applied to it.

Change-Id: I57cbf4f49cc18d9f842377eb0db6b66b1999283c