Editing a link URL via Special:LinkEdit or its API equivalent now makes an edit to the Link: page
Furthermore diffs are now actually viewable.

MW requirement is bumped up from unsupported, old LTS 1.35.3 to current, supported LTS, which is 1.39, due to the RestrictionStorage use in SpecialLinkEdit.php.
You could probably use $title->isProtected() for older MWs but it's time to let go of the unsupported version(s).
This will require code cleanup in other files but that can and should and will need to be done as a separate commit.

For the curious souls out there: the link URL and *only* the link URL is stored in the Link: page.
Other link metadata resides in the link table instead.

Change-Id: Ibe4e69f38c9cb4a24cb94e271e8ed63100159291